move revoke badge to project verification form section in adminJS
