+++ src/server/adminJs/tabs/projectsTab.ts
@@ -12,7 +12,6 @@ import {
   ProjectUpdate,
   ProjStatus,
   ReviewStatus,
-  RevokeSteps,
 } from '../../../entities/project';
 import { canAccessProjectAction, ResourceActions } from '../adminJsPermissions';
 import {
@@ -179,60 +178,6 @@ export const addFeaturedProjectUpdate = async (
     }),
     notice: {
       message: `ProjectUpdates(s) successfully added to the featured Project list`,
-      type: 'success',
-    },
-  };
-};
-
-export const revokeGivbacksEligibility = async (
-  context: AdminJsContextInterface,
-  request: AdminJsRequestInterface,
-) => {
-  const { records, currentAdmin } = context;
-  try {
-    const projectIds = request?.query?.recordIds
-      ?.split(',')
-      ?.map(strId => Number(strId)) as number[];
-    const updateParams = { isGivbackEligible: false };
-    const projects = await Project.createQueryBuilder('project')
-      .update<Project>(Project, updateParams)
-      .where('project.id IN (:...ids)')
-      .setParameter('ids', projectIds)
-      .returning('*')
-      .updateEntity(true)
-      .execute();
-
-    for (const project of projects.raw) {
-      const projectWithAdmin = (await findProjectById(project.id)) as Project;
-      projectWithAdmin.verificationStatus = RevokeSteps.Revoked;
-      await projectWithAdmin.save();
-      await getNotificationAdapter().projectBadgeRevoked({
-        project: projectWithAdmin,
-      });
-      const verificationForm = await getVerificationFormByProjectId(project.id);
-      if (verificationForm) {
-        await makeFormDraft({
-          formId: verificationForm.id,
-          adminId: currentAdmin.id,
-        });
-      }
-    }
-    await Promise.all([
-      refreshUserProjectPowerView(),
-      refreshProjectPowerView(),
-      refreshProjectFuturePowerView(),
-    ]);
-  } catch (error) {
-    logger.error('revokeGivbacksEligibility() error', error);
-    throw error;
-  }
-  return {
-    redirectUrl: '/admin/resources/Project',
-    records: records.map(record => {
-      record.toJSON(context.currentAdmin);
-    }),
-    notice: {
-      message: 'Project(s) successfully revoked from Givbacks eligibility',
       type: 'success',
     },
   };
@@ -1307,19 +1252,6 @@ export const projectsTab = {
           ),
         handler: async (request, response, context) => {
           return verifyProjects(context, request, false);
-        },
-        component: false,
-      },
-      revokeGivbacksEligible: {
-        actionType: 'bulk',
-        isVisible: true,
-        isAccessible: ({ currentAdmin }) =>
-          canAccessProjectAction(
-            { currentAdmin },
-            ResourceActions.REVOKE_BADGE,
-          ),
-        handler: async (request, _response, context) => {
-          return revokeGivbacksEligibility(context, request);
         },
         component: false,
       },
